\name{RmiRdata}
\alias{gene1}
\alias{gene2}
\alias{gene3}
\alias{mir1}
\alias{mir2}
\alias{mir3}
\docType{data}
\title{Simple demostration data for the RmiR package}

\description{
  Gene expression and microRNA expression data from the same RNA in a time course experiment
} 

\usage{
  data(RmiR)
}


\seealso{
  \code{RmiR_dbconn},\code{RmiR}
}

\examples{
  data(RmiR)
}

\keyword{datasets}
